Relocate certs packages to top-level folders
