+++ pkgs/swarmauri_certs_local_ca/tests/perf/test_service_perf.py
@@ -1,0 +1,35 @@
+import asyncio
+import pytest
+from cryptography.hazmat.primitives import serialization
+from cryptography.hazmat.primitives.asymmetric import rsa
+
+from swarmauri_certs_local_ca import LocalCaCertService
+from swarmauri_core.crypto.types import ExportPolicy, KeyRef, KeyType, KeyUse
+
+
+def _keyref() -> KeyRef:
+    sk = rsa.generate_private_key(public_exponent=65537, key_size=2048)
+    pem = sk.private_bytes(
+        serialization.Encoding.PEM,
+        serialization.PrivateFormat.PKCS8,
+        serialization.NoEncryption(),
+    )
+    return KeyRef(
+        kid="perf",
+        version=1,
+        type=KeyType.RSA,
+        uses=(KeyUse.SIGN,),
+        export_policy=ExportPolicy.SECRET_WHEN_ALLOWED,
+        material=pem,
+    )
+
+
+@pytest.mark.perf
+def test_self_signed_perf(benchmark) -> None:
+    svc = LocalCaCertService()
+    key = _keyref()
+
+    async def _run() -> None:
+        await svc.create_self_signed(key, {"CN": "perf"})
+
+    benchmark(lambda: asyncio.run(_run()))
