+++ pkgs/swarmauri_certs_local_ca/tests/unit/test_rfc2986_csr.py
@@ -1,24 +1,25 @@
-"""RFC 8032 Ed25519 certificate tests."""
+"""RFC 2986 PKCS#10 CSR structure tests."""
 
 import pytest
+from cryptography import x509
 from cryptography.hazmat.primitives import serialization
-from cryptography.hazmat.primitives.asymmetric import ed25519
+from cryptography.hazmat.primitives.asymmetric import rsa
 
-from swarmauri_certservice_localca import LocalCaCertService
+from swarmauri_certs_local_ca import LocalCaCertService
 from swarmauri_core.crypto.types import ExportPolicy, KeyRef, KeyType, KeyUse
 
 
 def _key() -> KeyRef:
-    sk = ed25519.Ed25519PrivateKey.generate()
+    sk = rsa.generate_private_key(public_exponent=65537, key_size=2048)
     pem = sk.private_bytes(
         serialization.Encoding.PEM,
         serialization.PrivateFormat.PKCS8,
         serialization.NoEncryption(),
     )
     return KeyRef(
-        kid="ed",
+        kid="csr",
         version=1,
-        type=KeyType.ED25519,
+        type=KeyType.RSA,
         uses=(KeyUse.SIGN,),
         export_policy=ExportPolicy.SECRET_WHEN_ALLOWED,
         material=pem,
@@ -27,8 +28,9 @@ def _key() -> KeyRef:
 
 @pytest.mark.unit
 @pytest.mark.asyncio
-async def test_self_signed_ed25519() -> None:
+async def test_csr_subject() -> None:
     svc = LocalCaCertService()
     key = _key()
-    cert = await svc.create_self_signed(key, {"CN": "ed"})
-    assert b"BEGIN CERTIFICATE" in cert
+    csr_bytes = await svc.create_csr(key, {"CN": "example"})
+    csr = x509.load_pem_x509_csr(csr_bytes)
+    assert csr.subject.rfc4514_string() == "CN=example"
