+++ pkgs/swarmauri_certs_local_ca/tests/unit/test_rfc5480_ecdsa.py
@@ -1,25 +1,24 @@
-"""RFC 2986 PKCS#10 CSR structure tests."""
+"""RFC 5480 ECDSA P-256 certificate tests."""
 
 import pytest
-from cryptography import x509
 from cryptography.hazmat.primitives import serialization
-from cryptography.hazmat.primitives.asymmetric import rsa
+from cryptography.hazmat.primitives.asymmetric import ec
 
-from swarmauri_certservice_localca import LocalCaCertService
+from swarmauri_certs_local_ca import LocalCaCertService
 from swarmauri_core.crypto.types import ExportPolicy, KeyRef, KeyType, KeyUse
 
 
 def _key() -> KeyRef:
-    sk = rsa.generate_private_key(public_exponent=65537, key_size=2048)
+    sk = ec.generate_private_key(ec.SECP256R1())
     pem = sk.private_bytes(
         serialization.Encoding.PEM,
         serialization.PrivateFormat.PKCS8,
         serialization.NoEncryption(),
     )
     return KeyRef(
-        kid="csr",
+        kid="ec",
         version=1,
-        type=KeyType.RSA,
+        type=KeyType.EC,
         uses=(KeyUse.SIGN,),
         export_policy=ExportPolicy.SECRET_WHEN_ALLOWED,
         material=pem,
@@ -28,9 +27,8 @@ def _key() -> KeyRef:
 
 @pytest.mark.unit
 @pytest.mark.asyncio
-async def test_csr_subject() -> None:
+async def test_self_signed_ecdsa() -> None:
     svc = LocalCaCertService()
     key = _key()
-    csr_bytes = await svc.create_csr(key, {"CN": "example"})
-    csr = x509.load_pem_x509_csr(csr_bytes)
-    assert csr.subject.rfc4514_string() == "CN=example"
+    cert = await svc.create_self_signed(key, {"CN": "ecdsa"})
+    assert b"BEGIN CERTIFICATE" in cert
