+++ pkgs/swarmauri_certs_pkcs11/README.md
@@ -1,6 +1,6 @@
 ![Swamauri Logo](https://res.cloudinary.com/dbjmpekvl/image/upload/v1730099724/Swarmauri-logo-lockup-2048x757_hww01w.png)
 
-# Swarmauri PKCS#11 Certificate Service
+# Swarmauri Certs PKCS#11
 
 A PKCS#11-backed certificate service implementing `CertServiceBase`.
 It generates and verifies X.509 certificates using hardware security modules.
@@ -8,7 +8,7 @@ It generates and verifies X.509 certificates using hardware security modules.
 ## Installation
 
 ```bash
-pip install swarmauri_certservice_pkcs11
+pip install swarmauri_certs_pkcs11
 ```
 
 ## Entry Point
