+++ pkgs/swarmauri_certs_pkcs11/tests/test_pkcs11_certservice_basic.py
@@ -2,7 +2,7 @@ import pytest
 
 pytest.importorskip("pkcs11")
 
-from swarmauri_certservice_pkcs11 import Pkcs11CertService
+from swarmauri_certs_pkcs11 import Pkcs11CertService
 
 
 @pytest.mark.test
