+++ pkgs/swarmauri_certs_pkcs11/tests/test_pkcs11_certservice_functional.py
@@ -8,7 +8,7 @@ from cryptography.x509.oid import NameOID
 
 pytest.importorskip("pkcs11")
 
-from swarmauri_certservice_pkcs11 import Pkcs11CertService
+from swarmauri_certs_pkcs11 import Pkcs11CertService
 
 
 @pytest.mark.asyncio
